requirements.txt
flask>=3.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
vaderSentiment>=3.3.2
plotly>=5.18.0
gunicorn>=21.0.0
python-multipart>=0.0.6
google-generativeai>=0.8.0
groq>=0.4.0
pillow>=9.0.0
pytest>=7.0.0

